Refactored Education frontend components (#1606)
* Refactored Education frontend components

* Move class fields to module level

* Move function into module level

* Replace function

* Combine functions in Employment form, other misc refactoring

* Use object for lookup

* More refactoring, delete some CSS

* More CSS tweaks and refactoring

* Rename describe
