+++ static/js/util/date_test.js
@@ -2,19 +2,31 @@
 import { assert } from 'chai';
 import moment from 'moment';
 
-import { ifValidDate } from './date';
+import { ifValidDate, formatMonthDate } from './date';
 
-describe('ifValidDate', () => {
+describe('date utility functions', () => {
   let formatFunc = date => date.format('YYYY-MM-DD');
   let testFunc = ifValidDate('not valid', formatFunc);
 
-  it('should return the default value for an invalid date', () => {
-    let invalidDate = moment('1978-13-39');
-    assert.equal('not valid', testFunc(invalidDate));
+  describe('ifValidDate', () => {
+    it('should return the default value for an invalid date', () => {
+      let invalidDate = moment('1978-13-39');
+      assert.equal('not valid', testFunc(invalidDate));
+    });
+
+    it('should return fn(date) for a valid date', () => {
+      let validDate = moment('1949-06-02');
+      assert.equal('1949-06-02', testFunc(validDate));
+    });
   });
 
-  it('should return fn(date) for a valid date', () => {
-    let validDate = moment('1949-06-02');
-    assert.equal('1949-06-02', testFunc(validDate));
+  describe('formatMonthDate', () => {
+    it('should format date as month/year', () => {
+      assert.equal('06/1949', formatMonthDate('1949-06-02'));
+    });
+
+    it('should return an empty string for a nil value', () => {
+      assert.equal('', formatMonthDate(null));
+    });
   });
 });
