+++ static/scss/dashboard.scss
@@ -390,80 +390,6 @@
     font-size: 12px;
   }
 
-  .profile-tab-card {
-    width: 100%;
-    position: relative;
-    overflow: visible;
-    margin-bottom: 35px;
-    padding-bottom: 25px;
-    min-height: 0;
-
-    .education-add-menu {
-      position: absolute;
-      top: 5%;
-      left: 75%;
-    }
-
-    .profile-tab-card-grid:first-child {
-      padding-top: 15px;
-    }
-
-    .profile-tab-card-grid.user-page {
-      .school-type {
-        font-size: 12px;
-        font-weight: 400;
-        text-transform: lowercase;
-      }
-
-      .mdl-cell {
-        margin: auto;
-      }
-    }
-
-    .profile-tab-card-grid {
-      margin: 0;
-      padding: 6px 15px;
-      $switchColor: #00a79d;
-
-      .profile-card-title {
-        font-weight: 600;
-        font-size: 19px;
-      }
-
-      .profile-row-icons {
-        text-align: right;
-        color: #898989;
-      }
-
-      .profile-row-name {
-        font-size: 16px;
-        font-weight: 700;
-      }
-
-      .profile-row-date-range {
-        font-size: 16px;
-        font-weight: 300;
-      }
-
-      .mdl-switch.is-checked {
-        .mdl-switch__track {
-          background-color: transparentize($switchColor, 0.5) !important;
-        }
-
-        .mdl-switch__thumb {
-          background-color: $switchColor !important;
-        }
-      }
-    }
-
-    .profile-add-button {
-      background-color: #0cbd8a !important;
-      bottom: -30px;
-      position: absolute;
-      right: 45px;
-    }
-  }
-
   .profile-tab-card-greyed {
     min-height: 70px !important;
     padding-bottom: 0;
